Totally changed a bogus sort to stooge
